Fix bug where boolean values are decoded incorrectly
TypeDecoder.decode() ignores the offset value specified for Boolean
types. This leads to incorrect decoding by the TypeDecoder.decodeBool() function,
which attempts to resolve the entire HEX input string instead of the portion of the
string identifying the boolean value.

In the HEX input given below for instance, where the
second value is the desired Boolean to decode, TextDecoder.decode
attempts to resolve the entire input as a Boolean value instead of the
64 characters starting at offset 64.

e.g. See TypeDecoderTest.testBooDecodeGivenOffset() for details.
0000000000000000000000000000000000000000000000007fffffffffffffff
0000000000000000000000000000000000000000000000000000000000000000
0000000000000000000000000000000000000000000000007fffffffffffffff